# docs/openapi.yaml
openapi: 3.1.0
info:
  title: API сервиса аутентификации
  version: 1.0.0
  summary: Аутентификация и вспомогательное API сервиса auth-service
  description: |
    Канонический API-контракт для текущего backend-сервиса.

    Модель аутентификации:
    - Access token возвращается в теле ответа как `accessToken`.
    - Refresh token хранится в HttpOnly cookie `refresh_token`.
    - Идентификатор refresh-сессии хранится в HttpOnly cookie `session_id`.
    - Для cookie-аутентифицированных эндпоинтов требуется CSRF-защита. Клиент должен сначала вызвать
      `GET /api/v1/auth/csrf`, затем отправить полученный токен в заголовке `X-XSRF-TOKEN` вместе
      с cookie `XSRF-TOKEN`.
    - `X-Refresh-Attempt-Id` настоятельно рекомендуется для `/api/v1/auth/refresh`. Повторное
      использование того же значения при сетевом retry позволяет backend безопасно повторить
      результат того же refresh-запроса.

    Политика Origin для браузерных запросов:
    - `allowedOrigins` в настройках auth client является источником истины для CORS.
    - Браузерные запросы с заголовком `Origin` разрешены только если origin принадлежит включенному auth client.
    - Запросы регистрации и входа дополнительно требуют, чтобы `Origin` принадлежал запрошенному `clientId`.
    - Refresh-запросы требуют, чтобы `Origin` принадлежал client, сохраненному в refresh-сессии.
    - Bearer-запросы требуют, чтобы `Origin` принадлежал claim `client_id` в access token.
...
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Аутентификация]
      summary: Зарегистрировать пользователя
      description: |
        Создает новую учетную запись пользователя, возвращает access token и выставляет refresh cookies.

        Требования по CSRF:
        - `XSRF-TOKEN` cookie
        - заголовок `X-XSRF-TOKEN` с соответствующим CSRF-токеном
      operationId: register
      parameters:
        - $ref: '#/components/parameters/UserAgentHeader'
        - $ref: '#/components/parameters/OriginHeader'
        - $ref: '#/components/parameters/XsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Регистрация выполнена успешно
          headers:
            Set-Cookie:
              $ref: '#/components/headers/AuthCookies'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Ошибка валидации запроса либо неизвестный/отключенный auth client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: CSRF-токен отсутствует/невалиден или Origin запроса не разрешен для запрошенного auth client.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Username или email уже существуют
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/auth/login:
    post:
      tags: [Аутентификация]
      summary: Выполнить вход пользователя
      description: |
        Аутентифицирует пользователя, возвращает access token и выставляет refresh cookies.

        Требования по CSRF:
        - `XSRF-TOKEN` cookie
        - заголовок `X-XSRF-TOKEN`
      operationId: login
      parameters:
        - $ref: '#/components/parameters/UserAgentHeader'
        - $ref: '#/components/parameters/OriginHeader'
        - $ref: '#/components/parameters/XsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Вход выполнен успешно
          headers:
            Set-Cookie:
              $ref: '#/components/headers/AuthCookies'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Ошибка валидации запроса либо неизвестный/отключенный auth client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Пользователь заблокирован, CSRF-токен отсутствует/невалиден или Origin запроса не разрешен для запрошенного auth client.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/ApiErrorResponse'
        '413':
          $ref: '#/components/responses/RequestBodyTooLarge'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/auth/refresh:
    post:
      tags: [Аутентификация]
      summary: Обновить access token
      description: |
        Ротирует refresh token и возвращает новый access token.

        Обязательные cookies:
        - `refresh_token`
        - `session_id`
        - `XSRF-TOKEN`

        Заголовки:
        - `X-XSRF-TOKEN` обязателен
        - `X-Refresh-Attempt-Id` настоятельно рекомендуется

        Использование `X-Refresh-Attempt-Id`:
        - Генерируйте новый UUID для каждой логической попытки refresh.
        - Повторно используйте то же значение только при повторе того же запроса из-за timeout или сетевого retry.
        - Если тот же старый refresh token повторно отправлен с другим attempt id в течение короткого grace-периода,
          backend вернет `409` с ошибкой о том, что refresh token уже обработан другим запросом.
      operationId: refresh
      parameters:
        - $ref: '#/components/parameters/UserAgentHeader'
        - $ref: '#/components/parameters/OriginHeader'
        - $ref: '#/components/parameters/XsrfHeader'
        - $ref: '#/components/parameters/RefreshAttemptIdHeader'
      responses:
        '200':
          description: Access token обновлен, refresh token ротирован
          headers:
            Set-Cookie:
              $ref: '#/components/headers/AuthCookies'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Auth client refresh-сессии отключен или больше недоступен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Невалидный refresh token, истекший token, обнаружен reuse refresh token или отсутствует refresh cookie
          headers:
            Set-Cookie:
              $ref: '#/components/headers/ClearedAuthCookies'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Пользователь заблокирован, CSRF-токен отсутствует/невалиден или Origin запроса не разрешен для auth client refresh-сессии.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/ClearedAuthCookies'
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Другой запрос уже ротировал тот же refresh token в пределах idempotency grace-периода
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/auth/logout:
    post:
      tags: [Аутентификация]
      summary: Выйти из текущей refresh-сессии
      description: |
        Отзывает текущий refresh token и очищает auth cookies.

        Требования по CSRF:
        - `XSRF-TOKEN` cookie
        - заголовок `X-XSRF-TOKEN`
      operationId: logout
      parameters:
        - $ref: '#/components/parameters/XsrfHeader'
      responses:
        '204':
          description: Выход выполнен успешно
          headers:
            Set-Cookie:
              $ref: '#/components/headers/ClearedAuthCookies'
        '403':
          description: Отсутствует или невалиден CSRF-токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout-all:
    post:
      tags: [Аутентификация]
      summary: Выйти из всех refresh-сессий
      description: Отзывает все refresh-сессии для аутентифицированного пользователя и очищает auth cookies на текущем клиенте.
      operationId: logoutAll
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Все refresh-сессии отозваны
          headers:
            Set-Cookie:
              $ref: '#/components/headers/ClearedAuthCookies'
        '401':
          description: Отсутствует или невалиден bearer token
  /api/v1/auth/logout-session/{sessionId}:
    post:
      tags: [Аутентификация]
      summary: Выйти из конкретной refresh-сессии
      description: |
        Отзывает все refresh token для выбранного session id аутентифицированного пользователя.
        Если запрошенный session id совпадает с текущей сессией, auth cookies также очищаются.
      operationId: logoutSession
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Сессия отозвана
          headers:
            Set-Cookie:
              $ref: '#/components/headers/ClearedAuthCookies'
        '401':
          description: Отсутствует или невалиден bearer token
  /api/v1/auth/password/change:
    post:
      tags: [Аутентификация]
      summary: Сменить пароль текущего пользователя
      description: |
        Меняет пароль аутентифицированного пользователя после проверки текущего пароля.
        Пользователь определяется только по bearer token.
        После успешной смены пароля backend отзывает остальные refresh-сессии пользователя,
        текущую сессию оставляет активной.
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Пароль изменен
        '400':
          description: Текущий пароль неверен, новый пароль совпадает с текущим или не проходит валидацию
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Отсутствует или невалиден bearer token
        '429':
          description: Превышен rate limit смены пароля
  /api/v1/auth/sessions:
    get:
      tags: [Аутентификация]
      summary: Получить список неотозванных refresh-сессий
      description: Истекшие, но еще не очищенные сессии могут присутствовать со статусом `EXPIRED`.
      operationId: listSessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Неотозванные сессии возвращены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
        '401':
          description: Отсутствует или невалиден bearer token
...
components:
  responses:
    RateLimitExceeded:
      description: Превышен rate limit для текущего scope запроса.
      headers:
        Retry-After:
          description: Рекомендуемая задержка перед повторной попыткой, в секундах.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: RATE_LIMIT_EXCEEDED
            message: Too many requests. Please try again later
    RequestBodyTooLarge:
      description: Тело login-запроса превышает допустимый размер.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: REQUEST_BODY_TOO_LARGE
            message: Request body is too large
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserAgentHeader:
      name: User-Agent
      in: header
      required: false
      schema:
        type: string
      description: Строка User-Agent клиента, сохраняемая в метаданных refresh-сессии.
    OriginHeader:
      name: Origin
      in: header
      required: false
      schema:
        type: string
        format: uri
      description: |
        Заголовок управляется браузером. Frontend-код не должен устанавливать его вручную.
        Если заголовок присутствует, он должен совпадать с одним из `allowedOrigins` найденного auth client.
        Небраузерные/server-to-server запросы обычно не отправляют этот заголовок.
    XsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
      description: CSRF-токен, возвращаемый `GET /api/v1/auth/csrf`.
    RefreshAttemptIdHeader:
      name: X-Refresh-Attempt-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: |
        Настоятельно рекомендуемый ключ идемпотентности для refresh-запроса.
        Повторно используйте то же значение только при retry того же логического refresh-запроса.
...
  headers:
...
    AuthCookies:
      description: Cookies refresh-сессии, выставляемые при login, register или refresh.
      schema:
        type: string
      example: refresh_token=abcdef123456; Path=/api/v1/auth; Max-Age=1209600; Expires=Sat, 25 Apr 2026 15:08:18 GMT; HttpOnly; Secure; SameSite=Lax, session_id=1f3a3b1e-2b31-4f78-8f2e-2ff0d4cb3fbc; Path=/api/v1/auth; Max-Age=1209600; Expires=Sat, 25 Apr 2026 15:08:18 GMT; HttpOnly; Secure; SameSite=Lax
    ClearedAuthCookies:
      description: Auth cookies, очищаемые backend-сервисом.
      schema:
        type: string
      example: refresh_token=; Path=/api/v1/auth; Max-Age=0; HttpOnly; Secure; SameSite=Lax, session_id=; Path=/api/v1/auth; Max-Age=0; HttpOnly; Secure; SameSite=Lax
  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        clientId:
          type: string
          description: Опциональный id auth client. По умолчанию используется `auth-service`.
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [username, password, email]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 8
          maxLength: 72
        email:
          type: string
          maxLength: 128
        clientId:
          type: string
          description: Опциональный id auth client. По умолчанию используется `auth-service`.
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
          maxLength: 72
          description: Новый пароль должен отличаться от текущего.
    TokenResponse:
      type: object
      additionalProperties: false
      required: [accessToken]
      properties:
        accessToken:
          type: string
          description: Подписанный JWT access token.
    SessionResponse:
      type: object
      additionalProperties: false
      required: [sessionId, clientId, userAgent, createdAt, expiresAt, revoked, status]
      properties:
        sessionId:
          type: string
        clientId:
          type: string
        ip:
          type: string
          nullable: true
        userAgent:
          type: string
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
        revoked:
          type: boolean
        revokedAt:
          type: string
          format: date-time
          nullable: true
        compromisedAt:
          type: string
          format: date-time
          nullable: true
        compromisedReason:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/SessionStatus'
    SessionStatus:
      type: string
      enum:
        - ACTIVE
        - REVOKED
        - EXPIRED
        - COMPROMISED
...
    ErrorResponse:
      type: object
      additionalProperties: true
      required: [timestamp, status, error, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        path:
          type: string
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - INVALID_REFRESH_TOKEN
            - REFRESH_TOKEN_EXPIRED
            - REFRESH_TOKEN_REUSE_DETECTED
            - INVALID_CREDENTIALS
            - USERNAME_ALREADY_TAKEN
            - EMAIL_ALREADY_REGISTERED
            - REFRESH_TOKEN_ALREADY_PROCESSED
            - INVALID_CLIENT
            - USER_NOT_FOUND
            - AUTH_CLIENT_ALREADY_EXISTS
            - AUTH_CLIENT_NOT_FOUND
            - RATE_LIMIT_EXCEEDED
            - ORIGIN_NOT_ALLOWED
            - INVALID_AUTH_CLIENT_ORIGIN
            - REQUEST_BODY_TOO_LARGE
            - USER_BANNED
            - INVALID_USER_BAN
            - USER_BAN_FORBIDDEN
            - USER_ALREADY_BANNED
        message:
          type: string
...